feat(ai): migrate AI Coach proxy to TypeScript; add tsconfig and CI smoke test
