+++ LiftTrackerAI/server/index.ts
@@ -1,71 +1,52 @@
-import express, { type Request, Response, NextFunction } from "express";
-import { registerRoutes } from "./routes";
-import { setupVite, serveStatic, log } from "./vite";
+import "dotenv/config";
+import express, { Request, Response } from "express";
+import cors from "cors";
+import Groq from "groq-sdk";
+
+type ChatMsg = { role: "system" | "user" | "assistant"; content: string };
 
 const app = express();
-app.use(express.json());
-app.use(express.urlencoded({ extended: false }));
+app.use(cors());
+app.use(express.json({ limit: "1mb" }));
 
-app.use((req, res, next) => {
-  const start = Date.now();
-  const path = req.path;
-  let capturedJsonResponse: Record<string, any> | undefined = undefined;
+const GROQ_API_KEY = process.env.GROQ_API_KEY;
+if (!GROQ_API_KEY) {
+  console.error("Missing GROQ_API_KEY");
+  process.exit(1);
+}
 
-  const originalResJson = res.json;
-  res.json = function (bodyJson, ...args) {
-    capturedJsonResponse = bodyJson;
-    return originalResJson.apply(res, [bodyJson, ...args]);
-  };
+const GROQ_MODEL = process.env.GROQ_MODEL || "llama-3.1-8b-instant";
+// 🔒 Lock here — only allow Llama 3.1 models
+const ALLOWED_MODELS = new Set<string>(["llama-3.1-8b-instant"]);
 
-  res.on("finish", () => {
-    const duration = Date.now() - start;
-    if (path.startsWith("/api")) {
-      let logLine = `${req.method} ${path} ${res.statusCode} in ${duration}ms`;
-      if (capturedJsonResponse) {
-        logLine += ` :: ${JSON.stringify(capturedJsonResponse)}`;
-      }
+const client = new Groq({ apiKey: GROQ_API_KEY });
 
-      if (logLine.length > 80) {
-        logLine = logLine.slice(0, 79) + "…";
-      }
-
-      log(logLine);
+app.post("/api/coach", async (req: Request, res: Response) => {
+  try {
+    const { messages } = req.body as { messages?: ChatMsg[] };
+    if (!Array.isArray(messages) || messages.length === 0) {
+      return res.status(400).json({ error: "Missing messages" });
     }
-  });
 
-  next();
+    if (!ALLOWED_MODELS.has(GROQ_MODEL)) {
+      return res.status(403).json({ error: "Model not allowed" });
+    }
+
+    const completion = await client.chat.completions.create({
+      model: GROQ_MODEL,
+      messages,
+      temperature: 0.3
+    });
+
+    const reply = completion?.choices?.[0]?.message?.content ?? "";
+    return res.json({ model: GROQ_MODEL, reply });
+  } catch (err: any) {
+    console.error("Coach error:", err?.message || err);
+    return res.status(500).json({ error: "AI Coach error" });
+  }
 });
 
-(async () => {
-  const server = await registerRoutes(app);
-
-  app.use((err: any, _req: Request, res: Response, _next: NextFunction) => {
-    const status = err.status || err.statusCode || 500;
-    const message = err.message || "Internal Server Error";
-
-    res.status(status).json({ message });
-    throw err;
-  });
-
-  // importantly only setup vite in development and after
-  // setting up all the other routes so the catch-all route
-  // doesn't interfere with the other routes
-  if (app.get("env") === "development") {
-    await setupVite(app, server);
-  } else {
-    serveStatic(app);
-  }
-
-  // ALWAYS serve the app on the port specified in the environment variable PORT
-  // Other ports are firewalled. Default to 5000 if not specified.
-  // this serves both the API and the client.
-  // It is the only port that is not firewalled.
-  const port = parseInt(process.env.PORT || '5000', 10);
-  server.listen({
-    port,
-    host: "0.0.0.0",
-    reusePort: true,
-  }, () => {
-    log(`serving on port ${port}`);
-  });
-})();
+const port = Number(process.env.PORT || 8080);
+app.listen(port, () => {
+  console.log(`AI Coach server (TS) on http://localhost:${port}`);
+});
