+++ tsconfig.json
@@ -1,0 +1,18 @@
+{
+  "compilerOptions": 
+  {
+    "target": "ES2020",
+    "lib": ["ES2020"],
+    "module": "CommonJS",
+    "moduleResolution": "Node",
+    "esModuleInterop": true,
+    "resolveJsonModule": true,
+    "strict": true,
+    "skipLibCheck": true,
+    "outDir": "dist",
+    "rootDir": ".",
+    "forceConsistentCasingInFileNames": true
+  },
+  "include": ["LiftTrackerAI/server/index.ts", "LiftTrackerAI/client/coach.ts"],
+  "exclude": ["node_modules", "dist"]
+}
